+++ App/Source/Primitive.cpp
@@ -5,10 +5,10 @@ using namespace BaseType;
 
 namespace Utilitary::Primitive
 {
-int8_t GetEdgeIndex(const Face& face, const IndexType v0Idx, const IndexType v1Idx)
+int GetEdgeIndex(const Face& face, const IndexType v0Idx, const IndexType v1Idx)
 {
 	// Initialize edge index to -1 (not found)
-	int8_t edgeIdx = -1;
+	int edgeIdx = -1;
 
 	// Lambda to find the edge index given a vertex and the current index
 	auto FindEdgeIndex = [&](const IndexType vertexIdx, const uint8_t idx)
@@ -16,11 +16,11 @@ int8_t GetEdgeIndex(const Face& face, const IndexType v0Idx, const IndexType v1I
 		IndexType prevVertexIdx = face.Vertices[IndexHelpers::Previous[idx]];
 		IndexType nextVertexIdx = face.Vertices[IndexHelpers::Next[idx]];
 		if(prevVertexIdx == vertexIdx)
-			return static_cast<int8_t>(IndexHelpers::Next[idx]);
+			return static_cast<int>(IndexHelpers::Next[idx]);
 		else if(nextVertexIdx == vertexIdx)
-			return static_cast<int8_t>(IndexHelpers::Previous[idx]);
+			return static_cast<int>(IndexHelpers::Previous[idx]);
 
-		return static_cast<int8_t>(-1);
+		return -1;
 	};
 
 	// Iterate through the vertices of the face to find the edge
